+++ internal/bundler/bundler_lower_test.go
@@ -1470,6 +1470,68 @@ func TestLowerStaticAsyncArrowSuperSetterES2016(t *testing.T) {
 	})
 }
 
+func TestLowerPrivateSuperES2022(t *testing.T) {
+	lower_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export { default as foo1 } from "./foo1"
+				export { default as foo2 } from "./foo2"
+				export { default as foo3 } from "./foo3"
+				export { default as foo4 } from "./foo4"
+				export { default as foo5 } from "./foo5"
+				export { default as foo6 } from "./foo6"
+				export { default as foo7 } from "./foo7"
+				export { default as foo8 } from "./foo8"
+			`,
+			"/foo1.js": `export default class extends x { #foo() { super.foo() } }`,
+			"/foo2.js": `export default class extends x { #foo() { super.foo++ } }`,
+			"/foo3.js": `export default class extends x { static #foo() { super.foo() } }`,
+			"/foo4.js": `export default class extends x { static #foo() { super.foo++ } }`,
+			"/foo5.js": `export default class extends x { #foo = () => { super.foo() } }`,
+			"/foo6.js": `export default class extends x { #foo = () => { super.foo++ } }`,
+			"/foo7.js": `export default class extends x { static #foo = () => { super.foo() } }`,
+			"/foo8.js": `export default class extends x { static #foo = () => { super.foo++ } }`,
+		},
+		entryPaths: []string{"/entry.js"},
+		options: config.Options{
+			Mode:                  config.ModeBundle,
+			UnsupportedJSFeatures: es(2022),
+			AbsOutputFile:         "/out.js",
+		},
+	})
+}
+
+func TestLowerPrivateSuperES2021(t *testing.T) {
+	lower_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				export { default as foo1 } from "./foo1"
+				export { default as foo2 } from "./foo2"
+				export { default as foo3 } from "./foo3"
+				export { default as foo4 } from "./foo4"
+				export { default as foo5 } from "./foo5"
+				export { default as foo6 } from "./foo6"
+				export { default as foo7 } from "./foo7"
+				export { default as foo8 } from "./foo8"
+			`,
+			"/foo1.js": `export default class extends x { #foo() { super.foo() } }`,
+			"/foo2.js": `export default class extends x { #foo() { super.foo++ } }`,
+			"/foo3.js": `export default class extends x { static #foo() { super.foo() } }`,
+			"/foo4.js": `export default class extends x { static #foo() { super.foo++ } }`,
+			"/foo5.js": `export default class extends x { #foo = () => { super.foo() } }`,
+			"/foo6.js": `export default class extends x { #foo = () => { super.foo++ } }`,
+			"/foo7.js": `export default class extends x { static #foo = () => { super.foo() } }`,
+			"/foo8.js": `export default class extends x { static #foo = () => { super.foo++ } }`,
+		},
+		entryPaths: []string{"/entry.js"},
+		options: config.Options{
+			Mode:                  config.ModeBundle,
+			UnsupportedJSFeatures: es(2021),
+			AbsOutputFile:         "/out.js",
+		},
+	})
+}
+
 func TestLowerClassField2020NoBundle(t *testing.T) {
 	lower_suite.expectBundled(t, bundled{
 		files: map[string]string{