+++ scripts/browser/browser-tests.js
@@ -186,37 +186,50 @@ let pages = {};
 
 for (let format of ['iife', 'esm']) {
   for (let min of [false, true]) {
-    for (let async of [false, true]) {
-      let code = `
-        window.testStart = function() {
-          esbuild.initialize({
-            wasmURL: '/esbuild.wasm',
-            worker: ${async},
-          }).then(() => {
-            return (${runAllTests})({ esbuild })
-          }).then(() => {
-            testDone()
-          }).catch(e => {
-            testFail('' + (e && e.stack || e))
-            testDone()
-          })
+    for (let worker of [false, true]) {
+      for (let module of [false, true]) {
+        let code = `
+          window.testStart = function() {
+            let promise = ${module}
+              ? esbuild.initialize({
+                  wasmURL: '/esbuild.wasm',
+                  worker: ${worker},
+                })
+              : fetch('/esbuild.wasm')
+                  .then(r => {
+                    if (!r.ok) throw new Error('Failed to download "/esbuild.wasm"')
+                    return r.arrayBuffer()
+                  })
+                  .then(ab => esbuild.initialize({
+                    wasmModule: new WebAssembly.Module(ab),
+                    worker: ${worker},
+                  }))
+            promise.then(() => {
+              return (${runAllTests})({ esbuild })
+            }).then(() => {
+              testDone()
+            }).catch(e => {
+              testFail('' + (e && e.stack || e))
+              testDone()
+            })
+          }
+        `;
+        let page;
+        if (format === 'esm') {
+          page = `
+            <script type="module">
+              import * as esbuild from '/esm/browser${min ? '.min' : ''}.js'
+              ${code}
+            </script>
+          `;
+        } else {
+          page = `
+            <script src="/lib/browser${min ? '.min' : ''}.js"></script>
+            <script>${code}</script>
+          `;
         }
-      `;
-      let page;
-      if (format === 'esm') {
-        page = `
-          <script type="module">
-            import * as esbuild from '/esm/browser${min ? '.min' : ''}.js'
-            ${code}
-          </script>
-        `;
-      } else {
-        page = `
-          <script src="/lib/browser${min ? '.min' : ''}.js"></script>
-          <script>${code}</script>
-        `;
+        pages[format + (min ? 'Min' : '') + (worker ? 'Worker' : '') + (module ? 'Module' : '')] = page;
       }
-      pages[format + (min ? 'Min' : '') + (async ? 'Async' : '')] = page;
     }
   }
 }